requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
captcha==0.6.0
celery==5.4.0
certifi==2024.8.30
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.0
cycler==0.12.1
Django==5.1.1
django-allauth==64.2.1
django-anymail==12.0
django-axes==6.5.1
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-filter==21.1
django-formtools==2.5.1
django-js-asset==2.2.0
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-private-storage==3.1.1
django-ratelimit==4.1.0
django-recaptcha==4.0.0
django-scheduler==0.10.1
django-storages==1.14.4
django-taggit==6.0.0
django-two-factor-auth==1.17.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
fonttools==4.53.1
gunicorn==23.0.0
icalendar==5.0.13
idna==3.8
kiwisolver==1.4.7
kombu==5.4.1
Markdown==3.7
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.24.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyparsing==3.1.4
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2024.2
qrcode==7.4.2
ratelimit==2.2.1
redis==5.0.8
reportlab==3.6.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
